pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^6.0.2"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
black = "^26.1.0"
pytest = "^9.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
